removed diffractometer routes to Diffractometer.py (as suggested by Mikel)
