+++ mxcube3/routes/Beamline.py
@@ -126,40 +126,6 @@ def getBeamInfo():
         resp.status_code = 200
         return resp
 
-
-@mxcube.route("/mxcube/api/v0.1/diffractometer/movables/state", methods=['GET'])
-def get_movables_state():
-    ret = {}
-
-    for movable in ['Phi', 'Focus', 'PhiZ', 'PhiY','Sampx', 'Sampy', 'Zoom']:
-        ret.update(Utils.get_movable_state_and_position(movable))
-
-    ret.update(Utils.get_light_state_and_intensity())
-
-    resp = jsonify(ret)
-    resp.status_code = 200
-    return resp
-
-
-@mxcube.route("/mxcube/api/v0.1/diffractometer/info", methods=['GET'])
-def get_diffractometer_info():
-    ret = {}
-
-    try:
-        ret['useSC'] = mxcube.diffractometer.use_sc  
-    except AttributeError:
-        ret['useSC'] = False # in case the diff does not have this implemented
-            
-    try:
-        ret['current_phase'] = mxcube.diffractometer.current_phase
-    except AttributeError:
-        ret['current_phase'] = 'None' # in case the diff does not have this implemented
-
-    resp = jsonify(ret)
-    resp.status_code = 200
-    return resp
-
-
 @mxcube.route("/mxcube/api/v0.1/sampleview/camera/info", methods=['GET'])
 def get_sample_video_dims():
     try:
